Election Poll
Python code to find data about election results. This data set is from a practice problem online that polled individuals on a daily basis, asking if an election was to be held on that day, which party would they vote for. The coding in this project practiced loading a csv file into the program to create a dataset, finding the mean, standard deviation, median and range of a particular party, graphing aggregate data for each political party into subplots, totaling votes and more.